pyproject.toml
[tool.poetry]
name = "hw27"
version = "0.1.0"
description = ""
authors = ["ipotemkin <91094458+ipotemkin@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.2"
Pillow = "^9.0.1"
psycopg2 = "^2.9.3"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.1.0"
python-dateutil = "^2.8.2"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.1.0"
pytest-cov = "^3.0.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
